requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-filter==21.1
django-heroku==0.3.1
django-paypal==2.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
filelock==3.7.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
oauthlib==3.2.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.1.0
